requirements.txt
requests
pdfplumber
python-docx
pandas
streamlit
xlsxwriter